requirements.txt
# KeyAuth - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi>=0.115.0
uvicorn>=0.30.0

# Database
tortoise-orm>=0.21.0
aiosqlite>=0.20.0
asyncpg>=0.29.0  # PostgreSQL driver for production

# Authentication & Security
bcrypt>=4.2.0
python-dotenv>=1.0.1

# Data Validation
pydantic>=2.9.0
email-validator>=2.2.0

# Machine Learning
scikit-learn>=1.5.0
numpy>=1.26.0
joblib>=1.4.0

# Email
aiosmtplib>=3.0.2

# Timezone data (required for Python 3.13 on Windows)
tzdata>=2024.1
